+++ ui/ui_mp_addressbook.c
@@ -90,7 +90,7 @@ void AddressBook_MenuInit( void )
 		s_addressbook_fields[i].visible_length	= 30;
 	//	strncpy(s_addressbook_fields[i].buffer, adr->string);
 		Q_strncpyz(s_addressbook_fields[i].buffer, adr->string, sizeof(s_addressbook_fields[i].buffer));
-		s_addressbook_fields[i].cursor = strlen(adr->string);
+		s_addressbook_fields[i].cursor = (int)strlen(adr->string);
 
 		Menu_AddItem(&s_addressbook_menu, &s_addressbook_fields[i]);
 	}
